Readme:
Synchronising users to Gitlab is now complete.
Included examples for excluding disabled LDAP user accounts.
User and group names to ignore are now case-insensitive.
Added potential feature: Specifying an attribute on the LDAP user in which this script could write back a user ID for each Gitlab instance.
Added potential feature: Likely the same as the above but for groups too. (Group renaming.)